pyServer/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
altgraph==0.17.2
async-timeout==3.0.1
attrs==19.3.0
backoff==1.10.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bs4==0.0.1
censys==0.0.8
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
colorlog==4.1.0
contextvars==2.4
cryptography==37.0.2
distro==1.7.0
dnspython==1.16.0
docker==4.2.0
eventlet==0.33.1
Flask==1.1.2
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
grpcio==1.46.3
grpcio-tools==1.46.3
HackRequests==1.2
httplib2==0.18.1
idna==2.9
idna-ssl==1.1.0
immutables==0.18
impacket==0.10.0
importlib-metadata==1.6.1
ipcalc==1.99.0
IPy==1.0
itsdangerous==1.1.0
Jinja2==2.11.2
ldap3==2.9.1
ldapdomaindump==0.9.3
lxml==4.9.0
MarkupSafe==1.1.1
more-itertools==8.4.0
msgpack==1.0.4
multidict==4.7.5
natsort==7.0.1
netaddr==0.7.19
ntlm-auth==1.5.0
numpy==1.18.5
packaging==20.4
pandas==1.0.5
paramiko==2.7.1
patsy==0.5.1
pexpect==4.8.0
pluggy==0.13.1
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.6.0
py==1.8.2
pyasn1==0.4.8
pycares==3.1.1
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.14.1
pyDes==2.0.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pysmb==1.2.1
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0
pyzmq==21.0.2
redis==3.4.1
requests~=2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
salt==3004.1
scapy==2.4.3
scipy==1.4.1
shodan==1.23.0
six==1.14.0
soupsieve==2.0.1
statsmodels==0.11.1
termcolor==1.1.0
tldextract==2.2.2
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3~=1.26.6
wcwidth==0.2.4
websocket-client==0.57.0
Werkzeug==1.0.1
XlsxWriter==1.2.8
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
